OEPS NEW APP/OEPS V2 Cusor/webapp_new/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
pyusb>=1.2.1
python-dotenv>=1.0.1
websockets>=12.0
numpy>=1.24.0
pandas>=2.2.0 